+++ CHANGELOG.md
@@ -8,6 +8,7 @@ This page lists the main changes made to Myokit in each release.
 - Deprecated
 - Removed
 - Fixed
+  - [#X](https://github.com/myokit/myokit/pull/X) Fixed bug in PatchMasterFile reader where data types other than int 16 were not detected.
 
 ## [1.37.4] - 2025-06-30
 - Added